wdt: fix coverity [1/1]
PD#SWPL-217775

Problem:
Potentially overflowing expression around operator [ * ],
which is evaluated using 4 byte arithmetic, and then used
in a context that expects an expression of 8 bytes type.

Solution:
Change long to int.

Verify:
SC2/AH212

(cherry picked from commit 7598a49c8381144da11c07bf1ee3cbbb03ae943b)

+++ drivers/watchdog/meson_gxbb_wdt.c
@@ -158,7 +158,7 @@ static int meson_gxbb_wdt_set_timeout(struct watchdog_device *wdt_dev,
 				      unsigned int timeout)
 {
 	struct meson_gxbb_wdt *data = watchdog_get_drvdata(wdt_dev);
-	unsigned long tcnt = timeout * 1000;
+	unsigned int tcnt = timeout * 1000;
 
 #if IS_ENABLED(CONFIG_AMLOGIC_DEBUG_IOTRACE)
 	if (wdt_debug)
@@ -194,7 +194,7 @@ static int meson_gxbb_wdt_set_pretimeout(struct watchdog_device *wdt_dev,
 					 unsigned int pretimeout)
 {
 	struct meson_gxbb_wdt *data = watchdog_get_drvdata(wdt_dev);
-	unsigned long tcnt = pretimeout * 1000;
+	unsigned int tcnt = pretimeout * 1000;
 
 #if IS_ENABLED(CONFIG_AMLOGIC_DEBUG_IOTRACE)
 	if (wdt_debug)
